homework2/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
flake8==3.4.1
idna==2.6
mccabe==0.6.1
py==1.4.34
pycodestyle==2.3.1
pycrypto==2.6.1
pyflakes==1.5.0
pytest==3.2.2
requests==2.18.4
urllib3==1.22
